+++ src/Krino/Krino.Domain.ConstructiveAdpositionalGrammar/Constructions/Rules/MaskRule.cs
@@ -1,6 +1,7 @@
 ﻿using Krino.Vertical.Utils.Enums;
 using Krino.Vertical.Utils.Rules;
 using System.Diagnostics;
+using System.Numerics;
 
 namespace Krino.Domain.ConstructiveAdpositionalGrammar.Constructions.Rules
 {
@@ -8,21 +9,21 @@ namespace Krino.Domain.ConstructiveAdpositionalGrammar.Constructions.Rules
     /// Evaluates true if the value contains the required bit mask.
     /// </summary>
     [DebuggerDisplay("{myMask}")]
-    public class MaskRule : IRule<ulong>
+    public class MaskRule : IRule<BigInteger>
     {
-        private ulong myMask;
+        private BigInteger myMask;
 
-        public MaskRule(ulong mask)
+        public MaskRule(BigInteger mask)
         {
             myMask = mask;
         }
 
-        public bool Evaluate(ulong value) => EnumBase.IsIn(myMask, value);
+        public bool Evaluate(BigInteger value) => EnumBase.IsIn(myMask, value);
 
 
-        public bool Equals(IRule<ulong> other) => other is MaskRule maskRule && myMask == maskRule.myMask;
+        public bool Equals(IRule<BigInteger> other) => other is MaskRule maskRule && myMask == maskRule.myMask;
 
 
-        public static MaskRule Is(ulong mask) => new MaskRule(mask);
+        public static MaskRule Is(BigInteger mask) => new MaskRule(mask);
     }
 }
